# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Discord Badge Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BadgeGuideStep {
  step: number;
  description: string;
  /** @nullable */
  url?: string | null;
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  /** One of: hypesquad, developer, staff, nitro, booster, special, legacy */
  category: string;
  /** One of: common, uncommon, rare, very_rare, legendary, legacy */
  rarity: string;
  /** Whether the badge is still obtainable today */
  obtainable: boolean;
  /** One of: instant, easy, medium, hard, unobtainable */
  difficulty: string;
  /**
     * Rough time to unlock (e.g. '2 minutes', '~1 month')
     * @nullable
     */
  timeEstimate?: string | null;
  /** Hex color for the badge theme */
  color: string;
  guide: BadgeGuideStep[];
  /**
     * JS snippet to paste in Discord's browser console, or null if not applicable
     * @nullable
     */
  consoleCommand?: string | null;
  /**
     * Short label for what the console command does
     * @nullable
     */
  consoleCommandLabel?: string | null;
  /**
     * Extra tips or caveats for unlocking
     * @nullable
     */
  tips?: string | null;
}

export type BadgeStatsByCategory = {[key: string]: number};

export type BadgeStatsByDifficulty = {[key: string]: number};

export interface BadgeStats {
  total: number;
  obtainable: number;
  legacy: number;
  byCategory: BadgeStatsByCategory;
  byDifficulty: BadgeStatsByDifficulty;
}

export interface BadgeCategory {
  id: string;
  label: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListBadgesParams = {
/**
 * Search query (name or description)
 */
q?: string;
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by whether the badge is currently obtainable
 */
obtainable?: ListBadgesObtainable;
/**
 * Filter by difficulty
 */
difficulty?: string;
};

export type ListBadgesObtainable = typeof ListBadgesObtainable[keyof typeof ListBadgesObtainable];


export const ListBadgesObtainable = {
  true: 'true',
  false: 'false',
} as const;