# lib/api-zod/src/generated/types/badge.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Discord Badge Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BadgeGuideStep } from './badgeGuideStep';

export interface Badge {
  id: string;
  name: string;
  description: string;
  /** One of: hypesquad, developer, staff, nitro, booster, special, legacy */
  category: string;
  /** One of: common, uncommon, rare, very_rare, legendary, legacy */
  rarity: string;
  /** Whether the badge is still obtainable today */
  obtainable: boolean;
  /** One of: instant, easy, medium, hard, unobtainable */
  difficulty: string;
  /**
     * Rough time to unlock (e.g. '2 minutes', '~1 month')
     * @nullable
     */
  timeEstimate?: string | null;
  /** Hex color for the badge theme */
  color: string;
  guide: BadgeGuideStep[];
  /**
     * JS snippet to paste in Discord's browser console, or null if not applicable
     * @nullable
     */
  consoleCommand?: string | null;
  /**
     * Short label for what the console command does
     * @nullable
     */
  consoleCommandLabel?: string | null;
  /**
     * Extra tips or caveats for unlocking
     * @nullable
     */
  tips?: string | null;
}