# lib/api-zod/src/generated/types/badgeGuideStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Discord Badge Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BadgeGuideStep {
  step: number;
  description: string;
  /** @nullable */
  url?: string | null;
}