# lib/api-zod/src/generated/types/badgeStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Discord Badge Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BadgeStatsByCategory } from './badgeStatsByCategory';
import type { BadgeStatsByDifficulty } from './badgeStatsByDifficulty';

export interface BadgeStats {
  total: number;
  obtainable: number;
  legacy: number;
  byCategory: BadgeStatsByCategory;
  byDifficulty: BadgeStatsByDifficulty;
}